+++ cli/tests/test_operations.rs
@@ -119,6 +119,18 @@ fn test_op_log() {
 }
 
 #[test]
+fn test_op_log_limit() {
+    let test_env = TestEnvironment::default();
+    test_env.jj_cmd_success(test_env.env_root(), &["init", "repo", "--git"]);
+    let repo_path = test_env.env_root().join("repo");
+
+    let stdout = test_env.jj_cmd_success(&repo_path, &["op", "log", "-Tdescription", "--limit=1"]);
+    insta::assert_snapshot!(stdout, @r###"
+    @  add workspace 'default'
+    "###);
+}
+
+#[test]
 fn test_op_log_template() {
     let test_env = TestEnvironment::default();
     test_env.jj_cmd_success(test_env.env_root(), &["init", "repo", "--git"]);